requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
Django==4.1.7
django-bootstrap4==22.3
exceptiongroup==1.1.0
fonttools==4.39.0
h11==0.14.0
idna==3.4
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
outcome==1.2.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pycparser==2.21
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
threadpoolctl==3.1.0
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
urllib3==1.26.15
wincertstore==0.2
wsproto==1.2.0
zipp==3.15.0
